refactor(frontend): change location of media state module
