+++ apps/frontend/app/routes/_dashboard.collections.$id._index.tsx
@@ -53,7 +53,7 @@ import {
 	useSearchParam,
 	useUserPreferences,
 } from "~/lib/hooks";
-import { useReviewEntity } from "~/lib/media";
+import { useReviewEntity } from "~/lib/state/media";
 import {
 	getAuthorizationHeader,
 	serverGqlService,
