+++ apps/frontend/app/routes/_dashboard.media.$action.$lot.tsx
@@ -76,7 +76,7 @@ import {
 	useUserDetails,
 	useUserPreferences,
 } from "~/lib/hooks";
-import { useMetadataProgressUpdate, useReviewEntity } from "~/lib/media";
+import { useMetadataProgressUpdate, useReviewEntity } from "~/lib/state/media";
 import {
 	getAuthorizationHeader,
 	serverGqlService,
