+++ apps/frontend/app/routes/_dashboard.media.people.item.$id._index.tsx
@@ -49,7 +49,7 @@ import {
 	ToggleMediaMonitorMenuItem,
 } from "~/components/media";
 import { useUserDetails, useUserPreferences } from "~/lib/hooks";
-import { useReviewEntity } from "~/lib/media";
+import { useReviewEntity } from "~/lib/state/media";
 import {
 	createToastHeaders,
 	getAuthorizationHeader,
